Linx: - chain consistency now checks for valid centromere traversal - extended cluster metrics for total range, deleted and chained

+++ sv-linx/src/main/java/com/hartwig/hmftools/linx/analysis/PairResolution.java
@@ -302,7 +302,7 @@ public class PairResolution
 
         boolean faceAway = (chainStart.position() < chainEnd.position()) == (chainStart.orientation() == 1);
 
-        int totalChainLength = chain.getLength(false);
+        long totalChainLength = chain.getLength(false);
         long syntheticLength = abs(chainStart.position() - chainEnd.position());
 
         ResolvedType resolvedType = faceAway ? ResolvedType.DEL : ResolvedType.DUP;
